+++ index.html
@@ -230,8 +230,7 @@
     </div>
   </template>
 
-  <script src="js/utils.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
