+++ js/utils.js
@@ -5,8 +5,6 @@ function getRandomPositiveFloat (min, max, digits = 1) {
   return result.toFixed(digits);
 }
 
-getRandomPositiveFloat;
-
 function getRandomPositiveInteger (min, max) {
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -16,12 +14,9 @@ function getRandomPositiveInteger (min, max) {
 
 const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
 
-
-getRandomArrayElement;
-
 // массив случайной длины из неповторяющихся значений. Взял с https://qna.habr.com/q/844269
-const createArr = ([...source], maxLength) => Array.from(
+const createRandomArray = ([...source], maxLength) => Array.from(
   { length: Math.min(source.length, 1 + Math.random() * maxLength | 0) },
   () => source.splice(Math.random() * source.length | 0, 1)[0]);
 
-createArr;
+export {getRandomPositiveFloat, getRandomPositiveInteger, getRandomArrayElement, createRandomArray};
